bsd-user: Block signals during sigreturn
This matches linux-user, and is needed to ensure we get the right
signal mask after a sigreturn. I'm not convinced passing ts->signal_mask
to sigprocmask inside process_pending_signals is correct, but it's what
linux-user does, and this is how linux-user ends up working rather than
leaving the previous set of taken signals blocked for the host, thereby
preventing back-to-back signals from being delivered.

In particular, this case fixes running bmake as an emulated process, as
it was previously getting stuck with SIGCHLD blocked for the host.